+++ tensorflow_quantum/core/ops/math_ops/tfq_simulate_1d_expectation.cc
@@ -127,9 +127,9 @@ class TfqSimulateMPS1DExpectationOp : public tensorflow::OpKernel {
 
     // Find largest circuit for tensor size padding and allocate
     // the output tensor.
-    int max_num_qubits = 0;
-    int min_num_qubits = 1 << 30;
-    for (const int num : num_qubits) {
+    uint64_t max_num_qubits = 0;
+    uint64_t min_num_qubits = 1 << 30;
+    for (const uint64_t num : num_qubits) {
       max_num_qubits = std::max(max_num_qubits, num);
       min_num_qubits = std::min(min_num_qubits, num);
     }
@@ -148,7 +148,7 @@ class TfqSimulateMPS1DExpectationOp : public tensorflow::OpKernel {
   int bond_dim_;
 
   void ComputeSmall(const std::vector<int>& num_qubits,
-                    const int max_num_qubits,
+                    const uint64_t max_num_qubits,
                     const std::vector<QsimCircuit>& unfused_circuits,
                     const std::vector<std::vector<PauliSum>>& pauli_sums,
                     tensorflow::OpKernelContext* context,
@@ -163,7 +163,7 @@ class TfqSimulateMPS1DExpectationOp : public tensorflow::OpKernel {
     auto DoWork = [&](int start, int end) {
       int old_batch_index = -2;
       int cur_batch_index = -1;
-      int largest_nq = 1;
+      uint64_t largest_nq = 1;
       int cur_op_index;
 
       // Note: ForArgs in MPSSimulator and MPSStateState are currently unused.
@@ -176,7 +176,7 @@ class TfqSimulateMPS1DExpectationOp : public tensorflow::OpKernel {
         cur_batch_index = i / output_dim_op_size;
         cur_op_index = i % output_dim_op_size;
 
-        const int nq = num_qubits[cur_batch_index];
+        const uint64_t nq = num_qubits[cur_batch_index];
 
         // (#679) Just ignore empty program
         auto unfused_gates = unfused_circuits[cur_batch_index].gates;
